# docs/swagger.yaml
basePath: /swagger/things-service
definitions:
...
  common.Page:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
...
  entity.DeviceInfo:
    properties:
      created_by:
        type: string
      created_time:
        type: string
      enabled:
        type: integer
      group_id:
        type: string
      id:
        type: string
      identifier:
        type: string
      json_data:
        type: string
      name:
        type: string
      parent_id:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      protocol_config:
        type: string
      status:
        type: integer
      tags:
        items:
          type: string
        type: array
      type:
        type: integer
      updated_by:
        type: string
      updated_time:
        type: string
    type: object
...
  model.Device_info:
    properties:
      created_by:
        description: created_by
        type: string
      created_time:
        description: created_time
        type: string
      enabled:
        description: enabled
        type: integer
      group_id:
        description: group_id
        type: string
      id:
        description: id
        type: string
      identifier:
        description: identifier
        type: string
      json_data:
        description: json_data
        type: string
      name:
        description: name
        type: string
      parent_id:
        description: parent_id
        type: string
      product_id:
        description: product_id
        type: string
      product_name:
        description: product_name
        type: string
      protocol_config:
        description: protocol_config
        type: string
      status:
        description: status
        type: integer
      type:
        description: type
        type: integer
      updated_by:
        description: updated_by
        type: string
      updated_time:
        description: updated_time
        type: string
    type: object
  model.Device_mirror:
    properties:
      id:
        description: 唯一标识
        type: string
      json_data:
        description: 当前数据json
        type: string
      updated_time:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: things-service API 文档.
  title: things-service RESTful API
  version: "1.0"
paths:
...
  /device-info:
    get:
      description: 查询所有实例, 可设置order, 以及查询条件等，例如 status=1, name=$like.%CAM% 等
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: protocol_config
        in: query
        name: protocol_config
        type: string
      - description: enabled
        in: query
        name: enabled
        type: string
      - description: identifier
        in: query
        name: identifier
        type: string
      - description: product_name
        in: query
        name: product_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Device_info'
                  type: array
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 查询所有实例
      tags:
      - Device_info
  /device-info/fuzzy-search-device:
    get:
      description: 根据标签、输入字符串模糊搜索查询设备(从标签的value 和名字中搜索), 如果要根据产品信息，则product_id 或 product_name
        要设置一个
      parameters:
      - description: tags标签，key:value,key:value形式
        in: query
        name: tags
        type: string
      - description: 模糊搜索字符串
        in: query
        name: q
        type: string
      - description: page
        in: query
        name: _page
        type: integer
      - description: pageSize
        in: query
        name: _page_size
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: product_name
        in: query
        name: product_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/entity.DeviceInfo'
                        type: array
                    type: object
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 模糊搜索查询设备
      tags:
      - Device_info
  /device-info/page:
    get:
      description: 分页查询所有实例, 可设置page, page_size, order, 以及查询条件等，例如 status=1, name=$like.%CAM%
        等
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: parent_id
        in: query
        name: parent_id
        type: string
      - description: group_id
        in: query
        name: group_id
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: protocol_config
        in: query
        name: protocol_config
        type: string
      - description: enabled
        in: query
        name: enabled
        type: string
      - description: identifier
        in: query
        name: identifier
        type: string
      - description: product_name
        in: query
        name: product_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/entity.DeviceInfo'
                        type: array
                    type: object
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 分页查询所有实例
      tags:
      - Device_info
  /device-info/search-device:
    get:
      description: 根据标签查询设备, 如果要根据产品信息，则product_id 或 product_name 要设置一个
      parameters:
      - description: tags标签，key:value,key:value形式
        in: query
        name: tags
        type: string
      - description: page
        in: query
        name: _page
        type: integer
      - description: pageSize
        in: query
        name: _page_size
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: identifier
        in: query
        name: identifier
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: product_id
        in: query
        name: product_id
        type: string
      - description: product_name
        in: query
        name: product_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例的数组
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/common.Page'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/entity.DeviceInfo'
                        type: array
                    type: object
              type: object
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 根据标签查询设备
      tags:
      - Device_info
...
  /device-mirror/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Device_mirror'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - Device_mirror
  /device-mirror/{identifier}:
    get:
      consumes:
      - application/json
      description: 获取实时device-mirror
      parameters:
      - description: identifier
        in: query
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功code和成功信息
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 错误code和错误信息
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取实时device-mirror
      tags:
      - Device_mirror
...
swagger: "2.0"